@@ -1,19 +1,18 @@
 {
   "name": "topify",
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "postinstall": "husky",
     "dev": "vite --port ${PORT:-5000}",
-    "build": "tsc && vite build",
-    "serve": "vite preview",
-    "lint:fix": "eslint ./src --ext .jsx,.js,.ts,.tsx --quiet --fix --ignore-path ./.gitignore",
-    "lint:format": "prettier  --loglevel warn --write \"./**/*.{js,jsx,ts,tsx,css,md,json}\" ",
+    "build": "tsc -b && vite build",
+    "preview": "vite preview",
+    "lint:fix": "eslint --quiet --fix",
+    "lint:format": "prettier  --log-level warn --write \"./**/*.{js,jsx,ts,tsx,css,md,json}\" ",
     "lint": "yarn lint:format && yarn lint:fix ",
     "type-check": "tsc",
     "upgrade:yarn": "yarn set version latest",
     "upgrade": "yarn upgrade-interactive ; yarn upgrade:yarn",
-    "vdev": "vercel dev",
-    "dev:vercel": "vercel dev --local-config ./vercel-dev.json",
     "dev:spotify": "nodemon --exec 'esr --cache' --ext 'ts' lib/spotify/test.ts",
     "check": "yarn lint && yarn type-check"
   },
@@ -27,29 +26,31 @@
     "react-router-dom": "^7.0.1"
   },
   "devDependencies": {
-    "@types/node": "^18.7.1",
-    "@types/react": "^18.0.17",
-    "@types/react-dom": "^18.0.6",
-    "@typescript-eslint/eslint-plugin": "^5.33.0",
-    "@typescript-eslint/parser": "^5.33.0",
-    "@vercel/node": "^2.5.7",
-    "@vitejs/plugin-react": "^2.0.0",
-    "esbuild-runner": "^2.2.1",
-    "eslint": "^8.21.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.30.1",
-    "eslint-plugin-simple-import-sort": "^7.0.0",
+    "@eslint/js": "^9.15.0",
+    "@types/node": "^22.10.1",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@vercel/node": "^3.2.27",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "esbuild-runner": "^2.2.2",
+    "eslint": "^9.16.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "eslint-plugin-react-refresh": "^0.4.14",
+    "eslint-plugin-simple-import-sort": "^12.1.1",
+    "globals": "^15.12.0",
     "husky": "^9.1.7",
-    "nodemon": "^2.0.19",
-    "prettier": "^2.7.1",
-    "rollup-plugin-visualizer": "^5.7.1",
-    "typescript": "^4.7.4",
-    "vercel": "^27.4.0",
-    "vite": "^3.0.5",
-    "vite-tsconfig-paths": "^3.5.0"
+    "nodemon": "^3.1.7",
+    "prettier": "^3.4.1",
+    "rollup-plugin-visualizer": "^5.12.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.15.0",
+    "vercel": "^39.1.2",
+    "vite": "^6.0.1",
+    "vite-plugin-vercel": "^9.0.3"
   },
   "license": "MIT",
   "packageManager": "yarn@4.5.3"
