@@ -1,17 +1,14 @@
+import react from '@vitejs/plugin-react-swc';
+import path from 'path';
 import { defineConfig } from 'vite';
+import vercel from 'vite-plugin-vercel';
 
-import react from '@vitejs/plugin-react';
-import tsconfigPaths from 'vite-tsconfig-paths';
-
-import { visualizer } from 'rollup-plugin-visualizer';
-
+// https://vite.dev/config/
 export default defineConfig({
-  plugins: [react(), tsconfigPaths(), visualizer()],
-  build: {
-    sourcemap: true,
+  plugins: [react(), vercel()],
+  resolve: {
+    alias: {
+      '~': path.resolve(__dirname, './src'),
+    },
   },
-  css: {
-    devSourcemap: true,
-  },
-  envPrefix: 'PUBLIC_',
 });
